Changed name form data/dummy.js to data/navbar-links, delete unused graphics and data.
